refactor: 공통 미리보기 별도 컴포넌트 추가
